Check for init system to get SSH cmd.
- Initalize distro with SSH connection info to set init system
- Set a timeout when executing test command in wait_for_ssh
  method.
- Wrap reboot command with sudo by default.
- Attempts in SSH connection methods count down not up.

+++ ipa/ipa_suse.py
@@ -6,13 +6,34 @@
 #
 # See LICENSE for license information.
 
+import ipa_utils
 from ipa_distro import Distro
+from ipa_exceptions import IpaDistroException
 
 
 class SUSE(Distro):
-    def __init__(self):
-        super(SUSE, self).__init__()
+    def __init__(self, instance_ip, ssh_private_key, ssh_user):
+        super(SUSE, self).__init__(instance_ip,
+                                   ssh_private_key,
+                                   ssh_user)
+
+    def _set_init_system(self, instance_ip, ssh_private_key, ssh_user):
+        """Determine the init system of distribution."""
+        out, err = ipa_utils.execute_ssh_command(
+            'ps -p 1 -o comm=',
+            instance_ip,
+            ssh_private_key,
+            ssh_user
+        )
+        self.init_system = out.strip()
 
     def get_stop_ssh_service_cmd(self):
         """Return command to stop SSH service for SUSE."""
-        return 'systemctl stop sshd'
+        if self.init_system == 'systemd':
+            return 'systemctl stop sshd'
+        elif self.init_system == 'init':
+            return 'rcsshd stop'
+        else:
+            raise IpaDistroException(
+                'The init system for SUSE distribution cannot be determined.'
+            )
